UI minor updates

+++ DiscountsAndPromotionsApp/UI/ProductCardScreen/ViewModel/PriceInfoViewModel.swift
@@ -2,19 +2,19 @@ import UIKit
 import Combine
 
 protocol PriceInfoViewViewModelProtocol {
-    var pricePublisher: AnyPublisher<Double, Never> { get }
-    var discountPricePublisher: AnyPublisher<Double, Never> { get }
+    var pricePublisher: AnyPublisher<Int, Never> { get }
+    var discountPricePublisher: AnyPublisher<Int, Never> { get }
     var addToFavorites: PassthroughSubject<Void, Never> { get }
     var isFavorite: Bool { get }
 
-    func updatePrice(_ price: Double)
-    func updateDiscountPrice(_ discountPrice: Double)
+    func updatePrice(_ price: Int)
+    func updateDiscountPrice(_ discountPrice: Int)
     func toggleFavorite()
 }
 
 class PriceInfoViewViewModel: PriceInfoViewViewModelProtocol {
-    @Published private (set) var price: Double = 0.0
-    @Published private (set) var discountPrice: Double = 0.0
+    @Published private (set) var price: Int = 0
+    @Published private (set) var discountPrice: Int = 0
     var addToFavorites = PassthroughSubject<Void, Never>()
     private let profileService: ProfileServiceProtocol
     private var product: Product
@@ -29,21 +29,21 @@ class PriceInfoViewViewModel: PriceInfoViewViewModelProtocol {
     }
 
     // Предоставляем @Published свойства как паблишеры
-    var pricePublisher: AnyPublisher<Double, Never> {
+    var pricePublisher: AnyPublisher<Int, Never> {
         $price.eraseToAnyPublisher()
     }
-    var discountPricePublisher: AnyPublisher<Double, Never> {
+    var discountPricePublisher: AnyPublisher<Int, Never> {
         $discountPrice.eraseToAnyPublisher()
     }
 
-    func updatePrice(_ price: Double) {
+    func updatePrice(_ price: Int) {
         DispatchQueue.main.async {
             print("Updating price to \(price)")
             self.price = price
         }
     }
 
-    func updateDiscountPrice(_ discountPrice: Double) {
+    func updateDiscountPrice(_ discountPrice: Int) {
         DispatchQueue.main.async {
             print("Updating discount price to \(discountPrice)")
             self.discountPrice = discountPrice
